backend/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
cssselect2==0.8.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
Js2Py==0.74
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyboard==0.13.5
llvmlite==0.44.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.6.0
MouseInfo==0.1.3
mpmath==1.3.0
narwhals==1.32.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
pipenv==2024.4.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.4
psutil==6.1.0
pure_eval==0.2.3
pyarrow==19.0.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydyf==0.11.0
PyGetWindow==0.0.9
Pygments==2.18.0
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyperclip==1.9.0
pyphen==0.17.2
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.14.0
rpds-py==0.24.0
semantic-version==2.10.0
setuptools==78.0.2
setuptools-rust==1.11.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.1
streamlit==1.44.0
svglib==1.5.1
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
weasyprint==65.0
webencodings==0.5.1
websockets==15.0.1
zopfli==0.2.3.post1
